Fix integers being passed as length to quiver3d.

+++ lib/mpl_toolkits/mplot3d/axes3d.py
@@ -2607,14 +2607,14 @@ pivot='tail', normalize=False, **kwargs)
             self.add_collection(linec)
             return linec
 
-        shaft_dt = np.array([0, length])
+        shaft_dt = np.array([0., length], dtype=float)
         arrow_dt = shaft_dt * arrow_length_ratio
 
         cbook._check_in_list(['tail', 'middle', 'tip'], pivot=pivot)
         if pivot == 'tail':
             shaft_dt -= length
         elif pivot == 'middle':
-            shaft_dt -= length/2.
+            shaft_dt -= length / 2
 
         XYZ = np.column_stack(input_args[:3])
         UVW = np.column_stack(input_args[3:argi]).astype(float)
